Add register_view func to register apps for convenience
Add a register_view decorator that can be applied to a view function
of an app. It adds it a global dictionary that can be imported and
served all at once. This is convenient b/c instead of having to
import all of the apps anywher you want to serve them (dunder main, tests, etc)
you can just import this dict.

Also added a templater convenience function to make it easier
to template a view