Fix bug where changing apps bleed into each other
Apps would bleed into each other when changing apps becuase the apps were defined once at definition time
And then at run time their different sections were appeneded to.
This results in the sections from each app just appearing below the previous app when you switch apps.
This is fixed by modify the register_view decorator to still define the links at definition time,
so that all apps have the same links (If the links were created at run time, you would have to run an
app before the link for that app was added to the links sections), but the apps themselves are created
fresh each time they are clicked on. This ensures that each app is new and each section of the app is
blank when it is created, so no bleed over effect is encountered.